+++ Samples/watchOS-Swift/watchOS-Swift WatchKit Extension/ExtensionDelegate.swift
@@ -12,7 +12,6 @@ class ExtensionDelegate: NSObject, WKExtensionDelegate {
                 return event
             }
             options.debug = true
-            options.logLevel = SentryLogLevel.verbose
             options.attachStacktrace = true
             options.sessionTrackingIntervalMillis = 5_000
         }