+++ Sources/Sentry/SentryLog.m
@@ -6,28 +6,25 @@ NS_ASSUME_NONNULL_BEGIN
 
 @implementation SentryLog
 
-+ (void)logWithMessage:(NSString *)message andLevel:(SentryLogLevel)level
+/**
+ * Enable per default to log initialization errors.
+ */
+static BOOL isDebug = YES;
+static SentryLevel diagnosticLevel = kSentryLevelError;
+
++ (void)configure:(BOOL)debug diagnosticLevel:(SentryLevel)level
 {
-    SentryLogLevel defaultLevel = kSentryLogLevelError;
-    if (SentrySDK.logLevel > 0) {
-        defaultLevel = SentrySDK.logLevel;
-    }
-    if (level <= defaultLevel && level != kSentryLogLevelNone) {
-        NSLog(@"Sentry - %@:: %@", [self.class logLevelToString:level], message);
+    isDebug = debug;
+    diagnosticLevel = level;
+}
+
++ (void)logWithMessage:(NSString *)message andLevel:(SentryLevel)level
+{
+    if (isDebug && level != kSentryLevelNone && level <= diagnosticLevel) {
+        NSLog(@"Sentry - %@:: %@", SentryLevelNames[level], message);
     }
 }
 
-+ (NSString *)logLevelToString:(SentryLogLevel)level
-{
-    switch (level) {
-    case kSentryLogLevelDebug:
-        return @"Debug";
-    case kSentryLogLevelVerbose:
-        return @"Verbose";
-    default:
-        return @"Error";
-    }
-}
 @end
 
 NS_ASSUME_NONNULL_END