+++ Tests/SentryTests/Helper/SentryLogTests.m
@@ -10,10 +10,14 @@
 
 - (void)testLogTypes
 {
-    [SentryLog logWithMessage:@"1" andLevel:kSentryLogLevelError];
-    [SentryLog logWithMessage:@"2" andLevel:kSentryLogLevelDebug];
-    [SentryLog logWithMessage:@"3" andLevel:kSentryLogLevelVerbose];
-    [SentryLog logWithMessage:@"4" andLevel:kSentryLogLevelNone];
+    [SentryLog configure:YES diagnosticLevel:kSentryLevelDebug];
+
+    [SentryLog logWithMessage:@"0" andLevel:kSentryLevelNone];
+    [SentryLog logWithMessage:@"1" andLevel:kSentryLevelDebug];
+    [SentryLog logWithMessage:@"2" andLevel:kSentryLevelInfo];
+    [SentryLog logWithMessage:@"3" andLevel:kSentryLevelWarning];
+    [SentryLog logWithMessage:@"4" andLevel:kSentryLevelError];
+    [SentryLog logWithMessage:@"5" andLevel:kSentryLevelFatal];
 }
 
 @end