+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - ref: Add read-only scope property to Hub #975
 
+Breaking Changes:
+
+- ref: Remove SentryException.userReported #974
+
 ## 7.0.0-alpha.0
 
 **Breaking Change**: This version introduces a change to the grouping of issues. The SDK now sets the `inApp`