+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## unreleased
 
+## 7.0.0-alpha.1
+
+Features and fixes:
+
 - ref: Add read-only scope property to Hub #975
 
 Breaking Changes: